+++ disconf-client/src/main/java/com/baidu/disconf/client/store/aspect/DisconfAspectJ.java
@@ -17,7 +17,7 @@ import com.baidu.disconf.client.common.annotations.DisconfItem;
 import com.baidu.disconf.client.config.DisClientConfig;
 import com.baidu.disconf.client.store.DisconfStoreProcessor;
 import com.baidu.disconf.client.store.DisconfStoreProcessorFactory;
-import com.baidu.disconf.client.utils.MethodUtils;
+import com.baidu.disconf.client.support.utils.MethodUtils;
 import com.baidu.disconf.core.common.constants.DisConfigTypeEnum;
 
 /**
