+++ disconf-client/src/main/java/com/baidu/disconf/client/support/utils/ConfigLoaderUtils.java
@@ -1,4 +1,4 @@
-package com.baidu.disconf.client.utils;
+package com.baidu.disconf.client.support.utils;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
